+++ packages/core/src/Viewer.tsx
@@ -39,11 +39,6 @@ export interface CharacterMap {
 }
 
 export interface ViewerProps {
-    // If you want to use an authorization header to access a PDF document from a protected server, then you can use
-    // `authorization: TOKEN_HERE`, for example:
-    // `authorization: 'Bearer ...'`
-    // Use `httpHeaders` option if you want to use other authorization server
-    authorization?: string;
     characterMap?: CharacterMap;
     // The default zoom level
     // If it's not set, the initial zoom level will be calculated based on the dimesion of page and the container width
@@ -61,6 +56,9 @@ export interface ViewerProps {
     renderError?: RenderError;
     renderPage?: RenderPage;
     renderLoader?(percentages: number): ReactElement;
+    // Indicate the cross-site requests should be made with credentials such as cookie and authorization headers.
+    // The default value is `false` 
+    withCredentials?: boolean;
     // The text selection mode
     selectionMode?: SelectionMode;
     onDocumentLoad?(e: DocumentLoadEvent): void;
@@ -69,11 +67,10 @@ export interface ViewerProps {
 }
 
 const Viewer: React.FC<ViewerProps> = ({
-    authorization = '',
     characterMap,
     defaultScale,
     fileUrl,
-    httpHeaders,
+    httpHeaders = {},
     initialPage = 0,
     localization,
     plugins = [],
@@ -81,6 +78,7 @@ const Viewer: React.FC<ViewerProps> = ({
     renderError,
     renderPage,
     renderLoader,
+    withCredentials = false,
     onDocumentLoad = () => {/**/},
     onPageChange = () => {/**/},
     onZoom = () => {/**/},
@@ -109,7 +107,6 @@ const Viewer: React.FC<ViewerProps> = ({
             <LocalizationProvider localization={localization}>
                 {(_) => ( // eslint-disable-line @typescript-eslint/no-unused-vars
                     <DocumentLoader
-                        authorization={authorization}
                         characterMap={characterMap}
                         file={file.data}
                         httpHeaders={httpHeaders}
@@ -142,6 +139,7 @@ const Viewer: React.FC<ViewerProps> = ({
                         )}
                         renderError={renderError}
                         renderLoader={renderLoader}
+                        withCredentials={withCredentials}
                     />
                 )}
             </LocalizationProvider>