+++ packages/core/src/index.d.ts
@@ -545,11 +545,6 @@ export interface CharacterMap {
 }
 
 export interface ViewerProps {
-    // If you want to use an authorization header to access a PDF document from a protected server, then you can use
-    // `authorization: TOKEN_HERE`, for example:
-    // `authorization: 'Bearer ...'`
-    // Use `httpHeaders` option if you want to use other authorization server
-    authorization?: string;
     characterMap?: CharacterMap;
     // The default zoom level
     // If it's not set, the initial zoom level will be calculated based on the dimesion of page and the container width
@@ -568,6 +563,9 @@ export interface ViewerProps {
     renderError?: RenderError;
     renderPage?: RenderPage;
     renderLoader?(percentages: number): ReactElement;
+    // Indicate the cross-site requests should be made with credentials such as cookie and authorization headers.
+    // The default value is `false`
+    withCredentials?: boolean;
     onDocumentLoad?(e: DocumentLoadEvent): void;
     onPageChange?(e: PageChangeEvent): void;
     onZoom?(e: ZoomEvent): void;